requirements.txt
Django >= 2.1, < 3.1
django-compressor
django-allauth == 0.42.0
django-angular >= 2.3
django-filer == 2.0.2
django-ipware
django-sass-processor
django-select2
django-filter == 2.4.0
django-rest-auth
django-polymorphic >= 2.1
django-admin-sortable2
django-fsm
django-fsm-admin
django-phonenumber-field == 4.0.0
django-post_office
django-cms >= 3.7.2, < 3.11
djangocms-bootstrap
djangocms-cascade >= 1.3.1
djangocms-text-ckeditor >= 3.9.1
libsass

djangorestframework==3.14.0
phonenumbers==8.13.30

#django-elasticsearch-dsl < 8
#elasticsearch-dsl >= 7, < 8
#djangoshop-paypal >= 1.2
#paypalrestsdk < 2

django-treebeard == 4.4

tox
pytest
pytest-django
beautifulsoup4
lxml
factory-boy
pytest-factoryboy
